Move interaction components into `components/interactions`
