+++ scripts/components/Interactions/HotspotNavButton.js
