+++ scripts/components/Interactions/NavigationButton.js
@@ -77,9 +77,58 @@ export const getLabelFromInteraction = (interaction) => {
   return {...interaction.label, labelText: interaction.labelText};
 };
 
+
+/**
+ * @typedef {{
+ *  sceneId: number;
+ *  interactionIndex: number;
+ *  isFocused: boolean;
+ *  buttonClasses: Array<string>;
+ *  icon: string;
+ *  showAsHotspot: boolean;
+ *  isHiddenBehindOverlay: boolean;
+ *  title: string;
+ *  label: string;
+ *  staticScene: boolean;
+ *  leftPosition: number;
+ *  topPosition: number;
+ *  type: string;
+ *  forceClickHandler: boolean;
+ *  nextFocus: string;
+ *  showHotspotOnHover: boolean;
+ *  isHotspotTabbable: boolean;
+ *  wrapperHeight: number;
+ *  rendered: boolean;
+ *  is3d: boolean;
+ *  clickHandler: () => void;
+ *  doubleClickHandler: () => void;
+ *  mouseDownHandler: (event: MouseEvent) => void;
+ *  onFocus: () => void;
+ *  onBlur: () => void;
+ *  onFocusedInteraction: () => void;
+ *  onMount: (wrapper: HTMLElement) => void;
+ *  onUnmount: (wrapper: HTMLElement) => void;
+ *  onUpdate: (wrapper: HTMLElement) => void;
+ * }} Props
+ */
+
+/**
+ * @typedef {{
+ *  isFocused: boolean;
+ *  expandButtonFocused: boolean;
+ *  innerButtonFocused: boolean;
+ *  isMouseOver: boolean;
+ * }} State
+ */
+
+
 export default class NavigationButton extends React.Component {
+  /**
+   * @param {Props} props 
+   */
   constructor(props) {
     super(props);
+    // this.props = {...this.props, ...props};
 
     this.navButtonWrapper = React.createRef();
     this.navButton = React.createRef();
@@ -152,17 +201,20 @@ export default class NavigationButton extends React.Component {
     }
   }
 
+  /**
+   * @param {Props} prevProps 
+   */
   componentDidUpdate(prevProps) {
     if (this.props.type && this.props.type === this.props.nextFocus && prevProps.nextFocus !== this.props.nextFocus) {
       this.skipFocus = true; // Prevent moving camera on next focus (makes for a better UX when using the mouse)
-      this.navButtonWrapper.current.focus({
+      this.navButtonWrapper.current && this.navButtonWrapper.current.focus({
         preventScroll: true
       });
     }
 
     if (this.props.isFocused && !prevProps.isFocused) {
       setTimeout(() => { // Note: Don't think the timeout is needed after rendering was fixed
-        this.navButtonWrapper.current.focus({
+        this.navButtonWrapper.current && this.navButtonWrapper.current.focus({
           preventScroll: true
         });
       }, 0);
@@ -298,7 +350,6 @@ export default class NavigationButton extends React.Component {
     const scene = this.context.params.scenes.find(
       (/** @type {Scene} */ scene) => scene.sceneId === this.props.sceneId,
     );
-    console.log(scene)
     const interaction = scene.interactions[this.props.interactionIndex];
     interaction.label.hotSpotSizeValues = widthX + "," + heightY;
   }
@@ -324,9 +375,11 @@ export default class NavigationButton extends React.Component {
     if (this.props.icon) {
       wrapperClasses.push(this.props.icon);
     }
-    if(this.props.showAsHotspot) {
+
+    if (this.props.is3d) {
       wrapperClasses = wrapperClasses.concat("nav-btn-hotspot render-in-3d");
     }
+
     if (this.state.isMouseOver) {
       wrapperClasses.push('hover');
     }
