+++ scripts/components/Interactions/NavigationButton.scss
