+++ scripts/components/Interactions/NavigationButtonLabel.scss
