+++ scripts/components/Interactions/OpenContent.js
@@ -1,0 +1,361 @@
+// @ts-check
+
+import React, { useCallback, useEffect, useRef } from "react";
+import "./OpenContent.scss";
+import { H5PContext } from "../../context/H5PContext";
+
+/**
+ * @typedef {{
+ *  sceneId: number;
+ *  interactionIndex: number;
+ *  handleFocus: () => void;
+ * }} Props
+ */
+
+export default class OpenContent extends React.Component {
+  /**
+   * @param {Props} props
+   */
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      anchorDrag: false,
+      canDrag: false,
+      camPosYaw: 0,
+      camPosPitch: 0,
+      startMousePos: 0,
+      startMidPoint: 0,
+      sizeWidth: 0,
+      sizeHeight: 0,
+    };
+
+    this.openContent = React.createRef();
+    this.openContentWrapper = React.createRef();
+  }
+
+  componentDidMount() {
+    const [sizeWidth, sizeHeight] = this.getHotspotValues();
+    this.setState({
+      sizeWidth,
+      sizeHeight,
+    });
+
+    if (this.props.onMount) {
+      // Let parent know this element should be added to the THREE world.
+      this.props.onMount(this.openContentWrapper.current);
+    }
+  }
+
+  getHotspotValues() {
+    const scene = this.context.params.scenes.find((scene) => {
+      return scene.sceneId === this.props.sceneId;
+    });
+    const interaction = scene.interactions[this.props.interactionIndex];
+
+    return interaction.label.hotSpotSizeValues
+      ? interaction.label.hotSpotSizeValues.split(",")
+      : [256, 128];
+  }
+
+  /**
+   * @param {number} widthX
+   * @param {number} heightY
+   */
+  setHotspotValues(widthX, heightY) {
+    const scene = this.context.params.scenes.find(
+      (/** @type {Scene} */ scene) => scene.sceneId === this.props.sceneId
+    );
+    const interaction = scene.interactions[this.props.interactionIndex];
+    interaction.label.hotSpotSizeValues = widthX + "," + heightY;
+  }
+  toggleDrag = (e) => {
+    const dragBool = !this.state.canDrag;
+    this.setState({
+      canDrag: dragBool,
+    });
+    if (!this.props.staticScene) {
+      //If we cant drag anymore, we start the rendering of the threesixty scene,
+      // we also set the camera position that is stored wen we start the hotspot scaling
+      if (!this.state.canDrag) {
+        this.context.threeSixty.startRendering();
+        this.context.threeSixty.setCameraPosition(
+          this.state.camPosYaw,
+          this.state.camPosPitch
+        );
+      } else {
+        //We store the current position, because we are technically still dragging the background around here
+        this.setState({
+          camPosYaw: this.context.threeSixty.getCurrentPosition().yaw,
+          camPosPitch: this.context.threeSixty.getCurrentPosition().pitch,
+        });
+        //We stop rendering the threesixty scene so it doesnt look like we are moving around
+        this.context.threeSixty.stopRendering();
+      }
+    }
+  };
+
+  onAnchorDragMouseDown = (e, horizontalDrag) => {
+    /*Based on the direction, we store the X or Y start position of the mouse,
+     and finds the center of the div, startMidPoint, which is needed for scaling from*/
+    this.setState({
+      anchorDrag: true,
+      startMousePos: horizontalDrag ? e.clientX : e.clientY,
+      startMidPoint: horizontalDrag
+        ? this.state.sizeWidth / 2
+        : this.state.sizeHeight / 2,
+    });
+  };
+  onMouseMove = (event, horizontalDrag) => {
+    //We record the currentMouseposition for everytime the mouse moves
+    const currentPosMouse = horizontalDrag ? event.clientX : event.clientY;
+
+    /*divStartWidth is the start mouse position subtracted by the midpoint, technically this
+    half the size of the actual div, this is used for keeping the original widtrh of the div
+    everytime we drag */
+    const divStartWidth = this.state.startMousePos - this.state.startMidPoint;
+
+    /* The final width is calculated by subtracting the position of the
+    mouse with the the divStartWidth, this is technically the offset between the
+    divStartWidth and the current mouse position. Since the div scales from the center,
+    we have to multiply the result by two*/
+
+    let finalValue = (currentPosMouse - divStartWidth) * 2;
+    if (finalValue > 64 && finalValue < 512) {
+      /*These values are used for inline styling in the div in the render loop,
+        updating the div dimensions when the mousemove event fires*/
+      horizontalDrag
+        ? this.setState({
+            sizeWidth: finalValue,
+          })
+        : this.setState({
+            sizeHeight: finalValue,
+          });
+    }
+  };
+
+  onAnchorDragMouseUp = (e, horizontalDrag) => {
+    let newSizeWidth = this.state.sizeWidth;
+    let newSizeHeight = this.state.sizeHeight;
+
+    this.setState({
+      anchorDrag: false,
+    });
+    //Used for writing the data into to editor, for them to persist into the viewer
+    this.setHotspotValues(newSizeWidth, newSizeHeight);
+  };
+
+  getStyle() {
+    const style = {};
+    if (this.props.topPosition !== undefined) {
+      style.top = this.props.topPosition + "%";
+    }
+
+    if (this.props.leftPosition !== undefined) {
+      style.left = this.props.leftPosition + "%";
+    }
+    return style;
+  }
+
+  getContentFromInteraction() {
+    const scene = this.context.params.scenes.find((scene) => {
+      return scene.sceneId === this.props.sceneId;
+    });
+    const interaction = scene.interactions[this.props.interactionIndex];
+    const library = interaction.action.library;
+    const machineName = H5P.libraryFromString(library).machineName;
+    if (machineName === "H5P.AdvancedText") {
+      return interaction.action.params.text;
+    } else if (machineName === "H5P.Image") {
+      const imgSrc = H5P.getPath(
+        interaction.action.params.file.path,
+        this.context.contentId
+      );
+      const image = `<img src=${imgSrc} alt=${interaction.action.params.alt}/>`;
+      return image;
+    } else {
+      return "";
+    }
+  }
+  onClick() {
+    const hasClickHandler = this.props.forceClickHandler
+      || !this.context.extras.isEditor;
+
+    if (hasClickHandler) {
+      this.props.clickHandler();
+
+      // Reset button focus state when changing scenes or opening content
+      this.setState({
+        innerButtonFocused: false
+      });
+    }
+  }
+
+  onDoubleClick() {
+    if (this.props.doubleClickHandler) {
+      this.props.doubleClickHandler();
+    }
+    this.setState({
+      isFocused: false,
+    });
+  }
+
+  onMouseDown(e) {
+    const hasMouseDownHandler = this.context.extras.isEditor
+      && this.props.mouseDownHandler;
+    if (hasMouseDownHandler) {
+      this.props.mouseDownHandler(e);
+    }
+  }
+  setFocus() {
+    const isFocusable = this.context.extras.isEditor
+      && this.openContentWrapper
+      && this.openContentWrapper.current;
+    if (isFocusable) {
+      this.openContentWrapper.current.focus({
+        preventScroll: true
+      });
+    }
+  }
+
+
+  handleFocus = (e) => {
+    if (this.context.extras.isEditor) {
+      if (this.openContentWrapper && this.openContentWrapper.current && this.openContentWrapper === e.target) {
+        this.openContentWrapper.current.focus({
+          preventScroll: true
+        });
+      }
+      return;
+    }
+
+    if (!this.context.extras.isEditor  && this.props.onFocus) {
+      if (this.skipFocus) {
+        this.skipFocus = false;
+      }
+      else {
+        this.props.onFocus();
+      }
+    }
+  }
+
+  render() {
+    let wrapperClasses = ["open-content-wrapper"];
+
+    if (this.state.isMouseOver) {
+      wrapperClasses.push("hover");
+    }
+
+    // only apply custom focus if we have children that are shown on focus
+    if (this.state.isFocused && this.props.children) {
+      wrapperClasses.push("focused");
+    }
+
+    // Add classname to current active element (wrapper, button or expand label button) so it can be shown on top
+    if (
+      (this.state.isFocused && this.props.children) ||
+      this.state.expandButtonFocused ||
+      this.state.innerButtonFocused
+    ) {
+      wrapperClasses.push("active-element");
+    }
+
+    const isWrapperTabbable = this.context.extras.isEditor;
+
+    const DragButton = (innerProps) => {
+      const hotspotBtnRef = useRef(null);
+
+      const mouseMoveHandler = (e) => {
+        this.onMouseMove(e, innerProps.horizontalDrag);
+      };
+      //Here we add a mouseup listener on the document so the user can release the mouse on anything on the document
+      const handleMouseDown = useCallback((e) => {
+        this.onAnchorDragMouseDown(e, innerProps.horizontalDrag);
+        this.toggleDrag();
+        document.addEventListener("mousemove", mouseMoveHandler);
+
+        document.addEventListener(
+          "mouseup",
+          () => {
+            document.removeEventListener("mousemove", mouseMoveHandler);
+            this.toggleDrag();
+            this.onAnchorDragMouseUp(e, innerProps.horizontalDrag);
+          },
+          { once: true }
+        );
+      }, []);
+
+      useEffect(() => {
+        /*In order to take control of the mousedown listener, we have to it when the component mount,
+       the reason for this is that we have to stop the propagation early on, since mousedown is already listened to by threesixty */
+        hotspotBtnRef.current.addEventListener("mousedown", (e) => {
+          e.stopPropagation();
+          handleMouseDown(e);
+        });
+      }, []);
+
+      return (
+        <button
+          className={
+            innerProps.horizontalDrag
+              ? "drag drag--horizontal"
+              : "drag drag--vertical"
+          }
+          ref={hotspotBtnRef}
+          tabIndex={this.props.tabIndexValue}
+          aria-label={
+            innerProps.horizontalDrag
+              ? this.context.l10n.hotspotDragHorizAlt
+              : this.context.l10n.hotspotDragVertiAlt
+          }
+        />
+      );
+    };
+
+    return (
+      <div
+        ref={this.openContentWrapper}
+        className={wrapperClasses.join(" ")}
+        style={this.getStyle()}
+        tabIndex={isWrapperTabbable ? 0 : undefined}
+        onFocus={this.handleFocus}
+      >
+        <div
+          className={`open-content ${
+            this.context.extras.isEditor ? "open-content--editor" : ""
+          }`}
+          ref={this.openContent}
+          aria-label={this.props.ariaLabel}
+          style={{
+            width: this.state.sizeWidth + "px",
+            height: this.state.sizeHeight + "px",
+          }}
+          tabIndex={this.props.tabIndexValue}
+          onDoubleClick={this.onDoubleClick.bind(this)}
+          onMouseDown={this.onMouseDown.bind(this)}
+          onMouseUp={this.setFocus.bind(this)}
+          onFocus={() => this.setState({ innerButtonFocused: true })}
+          onBlur={() => this.setState({ innerButtonFocused: false })}
+        >
+          <div
+            className={"inner-content"}
+            dangerouslySetInnerHTML={{
+              __html: this.getContentFromInteraction(),
+            }}
+          />
+          {this.context.extras.isEditor ? (
+            <>
+              <DragButton horizontalDrag={true} />
+              <DragButton horizontalDrag={false} />
+            </>
+          ) : (
+            ""
+          )}
+        </div>
+
+        {this.props.children}
+      </div>
+    );
+  }
+}
+OpenContent.contextType = H5PContext;
