+++ scripts/components/Interactions/OpenContent.scss
